refactor(project):extracted filesystem-store from core
