+++ packages/core/src/index.ts
@@ -1,7 +1,6 @@
 export * from './errors'
 export * from './models/physical-store'
 export * from './models/user'
-export * from './file-store'
 export * from './in-memory-store'
 export * from './store-manager'
 export * from './global-disposables'
