+++ packages/filesystem-store/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@furystack/filesystem-store",
+  "version": "1.0.0",
+  "description": "Simple File System store implementation for FuryStack",
+  "main": "dist/index.js",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "scripts": {},
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/furystack/furystack.git"
+  },
+  "keywords": [
+    "FuryStack",
+    "HTTP",
+    "REST",
+    "API",
+    "Actions"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "author": "gallay.lajos@gmail.com",
+  "license": "GPL-2.0",
+  "bugs": {
+    "url": "https://github.com/furystack/furystack/issues"
+  },
+  "homepage": "https://github.com/furystack/furystack",
+  "dependencies": {
+    "@furystack/core": "^8.1.0",
+    "@furystack/inject": "^4.0.27",
+    "@furystack/logging": "^2.0.34",
+    "@furystack/utils": "^1.1.17",
+    "semaphore-async-await": "^1.5.1",
+    "tslib": "^2.0.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^25.2.1",
+    "@types/redis": "^2.8.18",
+    "@types/uuid": "^8.0.0",
+    "uuid": "^8.0.0"
+  },
+  "typings": "./dist/index.d.ts",
+  "gitHead": "1045d854bfd8c475b7035471d130d401417a2321"
+}
