+++ packages/filesystem-store/src/index.ts
@@ -1,0 +1,2 @@
+export * from './filesystem-store'
+import './store-manager-extensions'
