+++ packages/filesystem-store/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "composite": true,
+    "rootDir": "./src"
+  },
+  "include": ["src"],
+  "references": [{ "path": "../core" }, { "path": "../inject" }, { "path": "../logging" }, { "path": "../utils" }]
+}
